Replaced helpers by user service
